feat: add isSafeFilename utility and integrate it into the Images2atlasWebpackPlugin
